Move FundSpecialTx function from rpcevo.cpp to specialtx_utils.cpp
Decoupling the RPC server errors from the function.
