+++ src/Makefile.am
@@ -190,6 +190,7 @@ BITCOIN_CORE_H = \
   evo/evonotificationinterface.h \
   evo/providertx.h \
   evo/specialtx_validation.h \
+  evo/specialtx_utils.h \
   flatdb.h \
   llmq/quorums_blockprocessor.h \
   llmq/quorums_commitment.h \
@@ -366,6 +367,7 @@ libbitcoin_server_a_SOURCES = \
   evo/mnauth.cpp \
   evo/providertx.cpp \
   evo/specialtx_validation.cpp \
+  evo/specialtx_utils.cpp \
   llmq/quorums_blockprocessor.cpp \
   llmq/quorums_commitment.cpp \
   llmq/quorums_connections.cpp \
